LayoutTests/http/tests/iframe-monitor/throttler-expected.txt

CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit: 10240 bytes, url=http://localhost:8080/iframe-monitor/resources/--eligible--/iframe.html
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit: 10240 bytes, url=http://localhost:8080/iframe-monitor/resources/--eligible--/iframe.html
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit: 10240 bytes, url=http://localhost:8080/iframe-monitor/resources/--eligible--/iframe.html
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit: 10240 bytes, url=http://localhost:8080/iframe-monitor/resources/--eligible--/iframe.html
CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit: 10240 bytes, url=http://localhost:8080/iframe-monitor/resources/--eligible--/iframe.html
CONSOLE MESSAGE: Frame's network usage exceeded the limit.
Test throttler prevents unloaded.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelector('iframe[name=frame1]').srcdoc is not ""
PASS document.querySelector('iframe[name=frame2]').srcdoc is not ""
PASS document.querySelector('iframe[name=frame3]').srcdoc is not ""
PASS document.querySelector('iframe[name=frame4]').srcdoc is not ""
PASS document.querySelector('iframe[name=frame5]').srcdoc is not ""
PASS result is true
PASS document.querySelector('iframe[name=frame6]').srcdoc is ""
PASS successfullyParsed is true

TEST COMPLETE
